+++ src/quodeq/services/suppression.py
@@ -23,51 +23,16 @@ from pathlib import Path
 from typing import Mapping
 
 from quodeq.config.paths import default_paths
+from quodeq.services.deleted import deleted_keys
+from quodeq.services.dismissed import dismissed_keys
+from quodeq.services.suppression_keys import (  # noqa: F401 — re-exported API
+    _coerce_line,
+    is_deleted,
+    is_dismissed,
+)
 from quodeq.shared.validation import validate_path_segment
 
 _TYPE_VIOLATION = "violation"
-
-
-def _coerce_line(line: object) -> int:
-    try:
-        return int(line)  # type: ignore[arg-type]
-    except (TypeError, ValueError):
-        return 0
-
-
-def is_dismissed(
-    dismissed: frozenset | set, *, req: str | None, principle: str | None = "",
-    file: str | None = "", line: object = 0,
-) -> bool:
-    """True when the dismiss store hides this finding.
-
-    A finding's dismiss identity is its ``req``, falling back to its principle
-    when it has none -- the same ``req || principle`` the UI stores
-    (buildDismissPayload). Every read side must apply the same fallback, or a
-    no-req finding disappears from the counters while its grade never moves.
-
-    For a no-req finding the assistant's draft/apply path records the key with
-    an empty req (``("", file, line)``, see tests/assistant/
-    test_dismiss_apply_e2e.py), and older stores may hold either form -- so
-    both are accepted.
-    """
-    if not dismissed:
-        return False
-    file_key = file or ""
-    line_key = _coerce_line(line)
-    if (req or principle or "", file_key, line_key) in dismissed:
-        return True
-    return not req and ("", file_key, line_key) in dismissed
-
-
-def is_deleted(
-    deleted: frozenset | set, *, dimension: str | None,
-    principle: str | None, file: str | None,
-) -> bool:
-    """True when the delete store hides this finding's whole principle+file."""
-    if not deleted:
-        return False
-    return (dimension or "", principle or "", file or "") in deleted
 
 
 @dataclass(frozen=True)
@@ -148,9 +113,6 @@ def project_suppressions(project_dir: Path) -> tuple[frozenset, frozenset]:
     memo keyed on anything coarser than the file contents would freeze the
     counts for the rest of the run.
     """
-    from quodeq.services.deleted import deleted_keys  # noqa: PLC0415
-    from quodeq.services.dismissed import dismissed_keys  # noqa: PLC0415
-
     return frozenset(dismissed_keys(project_dir)), frozenset(deleted_keys(project_dir))
